refactor: modify the structure
