+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from 'components/header';
+import Header from 'components/layout/header';
 
 import GlobalStyle from 'styles/globalStyle';
